requirements.txt
langchain
python-dotenv
langchain-community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve[all]
langchain-groq
SQLAlchemy
mysql-connector-python
huggingface_hub
duckduckgo-search
tavily-python